feat: ch5_parse_float_num

+++ examples/ch5_f32_as_u32.rs
@@ -11,7 +11,7 @@ fn main() {
     println!("{:032b}", frankentype);
 
     // 再将其转换为 f32 类型
-    let b: f32 = unsafe { std::mem::transmute(frankentype) };
+    let b: i32 = unsafe { std::mem::transmute(frankentype) };
     println!("{}", b);
     //单个字节内部的布局顺序的偏好叫做位编号或者位端序
 }
